security_dac: Restore original owner more often
Now that we know what label we should restore and we do have
reference counter to each seclabel, we restore the original
seclabel only after the last domain is torn down. Therefore, we
can safely try to restore labels even for RO or shared disks. The
reference counter will take the care of everything.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>