remote_driver: Implement virStreamSkip
Now that we have RPC wrappers over VIR_NET_STREAM_SKIP we can
start wiring them up. This commit wires up situation when a
client wants to send a hole to daemon.

To keep stream offsets synchronous, upon successful call on the
daemon skip the same hole in local part of the stream.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>