fdstream: Suppress use of IO helper for sparse streams
This is kind of a hacky approach to the following problem, but so
far I am unable to come up with anything better. On some
occasions (esp. when dealing with regular files) libvirt_iohelper
is spawned to prefetch data for us. We will then have a pipe then
for reading the data from it. This does not fit in our sparse
stream implementation as one simply doesn't lseek() over a pipe.
Until this is resolved, let's suppress use of the IO helper and
read data from FD directly.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>