fdstream: Suppress use of IO helper for sparse streams
NOT TO BE APPLIED UPSTREAM!

This is just a hack so that reviewers can try sparse streams out.
Problem is, if nonblocking streams are requested (which is case
of every single stream in the daemon), we spawn IO helper and
create a pipe to read data from it. But that means fd that
virFDStreamInData() sees is to our end of pipe, not the actual
file. Therefore any lseek() fails (one simply doesn't seek over a
pipe). Solution would be to bring some protocol to our
communication with the IO helper. But that's going to be yet
another N patches and this patch set is long enough as is. So
just for the sake of review, do this hack for now.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>