iohelper: Move runIO to runIOBasic
The runIO function is after all initialization (e.g. argument
parsing, file opening, ...) is done. It is the heart of the
iohelper. Unfortunately, approach implemented in next patches is
not easy to be combined with O_DIRECT. Therefore, rename runIO to
runIOBasic and create new runIO, which for now just wraps over
runIOBasic. But in not so distant future it will decide which IO
loop is going to be called.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>