qemuDomainDetachDeviceFlags: Pass live device XML to live detach code
The problem is this: when working on redirdev detach, I've
noticed that even though I've passed device alias in the input
device XML, it got transformed into inactive in
qemuDomainDetachDeviceLive() while in
qemuDomainDetachDeviceConfig() it was the active version of it.

Apparently, if you are doing:

virsh # detach-device --config --live domain device.xml

our code correctly detects that you want to detach the device
from both config and live domain definition. However, due to some
internal implementation we need to make a copy of the device that
user had provided. And to do that, we call
virDomainDeviceDefCopy(). This function basically formats the
device into XML and then parse it again. But there's a problem,
it's formatting the XML with VIR_DOMAIN_DEF_FORMAT_INACTIVE flag
which means that no live data are present in the copy. So we have
a (possibly live) device definition that we pass to code
detaching it from inactive XML, and inactive device definition
that we pass to code detaching it from active XML. This makes no
sense.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>