fdstream: Track formatted message
When using iohelper and sparse streams at once, we need a
protocol between iohelper and daemon to tell the other side to
represent received data in some fashion (e.g. whether it is raw
data, or a size of hole). In this patch, virFDStreamDrv tracks
whether the assumption from the first line stands.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>