iohelper_message: Implement formatted read
This is a bit tricky to grasp. But lets try anyway. So we read
data from a pipe. Now, the data are not raw file data as they
used to be. But they are a formatted message (encoded
virNetMessage in this case). So until we've read it whole, we
must: a) continue reading its remainder  b) if reading would
block and caller wants us to read nonblock, claim EAGAIN and
don't return any partially read data.
Now, another interesting thing may happen - we've read a big
message, decoded it but caller is reading just a small chunks. In
that case we must not read any new message but rather copy those
small chunks into caller's buffer until the whole message is
processed. Only after that we can read new message from the pipe
and process continues.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>