iohelper_message: Implement formatted write
This is nearly the same story as formatted read. With one
exception. Due to the recurring pattern of calling write
functions in our code:

  repeat {
    nwritten = write(dest, buf + offset, buflen - offset);
    offset += nwritten;
  }

we have to be cautious about the return value of iohelperWrite().
If we would return number of bytes partially written, we would be
called again with an offset shifted. But at the beginning of
iohelperWrite() we flush the output queue so the remaining data
is written then. Moreover, there's no way for us to determine
whether this is the case or we are called with fresh data,
completely unrelated to the first call.
Therefore, we must keep claiming EAGAIN, even though the message
is being partially written, and just on the last iteration claim
success and return the size of data we were requested to write in
the first place. Even if this means that in the last iteration
just one byte was written to make the message write complete.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>