Introduce virDomainMemoryBacking struct
This reflects our structure that we have in the XML. Moreover,
this structure is expected to grow in near future, therefore it
makes only a little sense to put the new items into
virDomainMemtune struct and make it bigger needlessly.

The parsing of <memoryBacking/> element and its children is put
into virDomainMemoryBackingParse() which for now does plain
parsing. This is because our code is not ready for
<memoryBacking/> being NULL. This will change soon.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>